+++ example/lib/json.dart
@@ -1,0 +1,41 @@
+final json = <String, dynamic>{
+  'squadName': 'Hidden',
+  'homeTown': 'Metro City',
+  'formed': 2016,
+  'secretBase': 'Super tower',
+  'active': true,
+  'members': [
+    <String, dynamic>{
+      'name': 'Molecule Man',
+      'age': 29,
+      'secretIdentity': 'Dan Jukes',
+      'powers': [
+        'Radiation resistance',
+        'Turning tiny',
+        'Radiation blast',
+      ],
+    },
+    <String, dynamic>{
+      'name': 'Madame Uppercut',
+      'age': 39,
+      'secretIdentity': 'Jane Wilson',
+      'powers': [
+        'Million tonne punch',
+        'Damage resistance',
+        'Superhuman reflexes',
+      ],
+    },
+    <String, dynamic>{
+      'name': 'Eternal Flame',
+      'age': 1000.00,
+      'secretIdentity': 'Unknown',
+      'powers': [
+        'Immortality',
+        'Heat Immunity',
+        'Inferno',
+        'Teleportation',
+        'Interdimensional travel',
+      ],
+    },
+  ],
+};