+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+* Added new widget: **CrJsonRecyclerWidget**
+* Added new widget: **CrJsonRecyclerSliver**
+* Updated README
+
 ## 0.9.1
 
 * Updated README